requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
anyio==4.9.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
et_xmlfile==2.0.0
feedparser==6.0.11
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.1
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
schedule==1.2.2
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
Telethon==1.38.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yadisk==3.2.0
yarl==1.18.3
